pyproject.toml
[project]
name = "finmas"
version = "0.1.0"
description = "Financial Analysis using Multi-Agent Systems"
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
authors = [
  { name = "Ivar Soares Urdalen" }
]
requires-python = ">=3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "alpaca-py>=0.42.0",
  "alpha-vantage>=3",
  "crewai>=0.140.0",
  "crewai-tools>=0.49.0",
  "datamule>=0.363",
  "datasets>=3.6.0",
  "diskcache>=5.6.3",
  "edgartools>=4.3.1",
  "feedparser>=6.0.11",
  "financedatabase>=2.3.1",
  "html-to-markdown>=1.4.0",
  "html2text>=2025.4.15",
  "huggingface-hub[cli]",
  "hvplot>=0.11.3",
  "ipywidgets>=8.1.7",
  "jupyterlab>=4.4.4",
  "jupyterlab-quarto>=0.3.5",
  "litellm==1.72.6",
  "llama-index>=0.12.46",
  "llama-index-embeddings-huggingface>=0.5.5",
  "llama-index-llms-groq>=0.3.2",
  "matplotlib>=3.10.3",
  "nbconvert>=7.16.6",
  "pandas>=2.3.0",
  "panel>=1.7.2",
  "plotly>=6.2.0",
  "python-dotenv>=1.1.1",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.7.0",
  "seaborn>=0.13.2",
  "sentence-transformers>=5.0.0",
  "ta>=0.11",
  "tabulate>=0.9",
  "tiingo>=0.16.1",
  "tiktoken>=0.9.0",
  "torch==2.7.1+cpu",
  "tqdm>=4.67.1",
  "transformers>=4.46.3",
  "typer>=0.16.0",
]

[dependency-groups]
dev = [
  "mypy>=1.16.1",
  "pandas-stubs>=2.3.0.250703",
]
docs = [
  "markdown-exec[ansi]>=1.11.0",
  "mkdocs>=1.6.1",
  "mkdocs-macros-plugin>=1.3.7",
  "mkdocs-material>=9.6.15",
  "mkdocs-mermaid2-plugin>=1.2.1",
  "pymdown-extensions>=10.16",
  "mdformat-black>=0.1.1",
  "mdformat-frontmatter>=2.0.8",
  "mdformat-gfm>=0.4.1",
  "mdformat-mkdocs>=4.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finmas"]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# Linting and formatting

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
ignore = [
    "E501",   # line-too-long
    "D100",   # missing-docstring-in-public-module
    "D101",   # missing-docstring-in-public-class
    "D104",   # missing-docstring-in-public-package
    "D105",   # missing-docstring-in-magic-method
    "D107",   # missing-docstring-in-init
    "D212",   # multi-line-summary-first-line
    "C408",   # unnecessary-collection-call
]
select = [
    "ARG001",  # unused-function-argument
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "E",       # pycodestyle-error
    "F",       # pyflakes
    "I",       # isort
    "UP",      # pyupgrade
    "W",       # pycodestyle-warning
    "D",       # pydocstyle
]
extend-select = ["D213"]
unfixable = [
  "F401", # F401 unused imports
]
[tool.ruff.lint.isort]
known-first-party = ["finmas", "tests"]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # unused-import
    "F403", # unused-wildcard-import
]
"tests/*" = [
    "ARG001", # unused-function-argument
    "D103", # missing-docstring-in-public-function
]
"**/crews/*" = [
    "D102", # undocumented-public-method
    "D103", # undocumented-public-function
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10
max-branches = 20
max-statements = 60
